+++ ext_localconf.php
@@ -21,12 +21,6 @@ $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['Tx_Themes_Domain_Repository_ThemeRepo
 $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tstemplate.php']['includeStaticTypoScriptSourcesAtEnd'][]
     = \KayStrobach\Themes\Hooks\T3libTstemplateIncludeStaticTypoScriptSourcesAtEndHook::class . '->main';
 
-/*
- * register hook to manipulate the template module
- */
-$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['typo3/template.php']['moduleBodyPostProcess'][]
-    = 'KayStrobach\\Themes\\Hooks\\TemplateModuleBodyPostProcessHook->main';
-
 /**
  * register signal to inject pagets without xclassing
  * Requires signal slot call from http://forge.typo3.org/issues/59703.