+++ ChangeLog.md
@@ -18,6 +18,7 @@
 *   [BREAKING] Remove logic for setup.txt/constants.txt/tsconfig.txt - only *.typoscript files will be recognized.
 *   [BREAKING] Remove constant view helper including cache logic - please pass your constants by plugin settings for achieving better performance.
 *   [TASK] Remove t3jquery.txt configuration.
+*   [TASK] Remove of the moduleBodyPostProcess Hook which is no longer available in TYPO3 9.5.